rename: cqrs패턴으로 group 디렉토리 구조 변경
