+++ src/main/java/com/fivemybab/ittabab/group/command/application/service/GroupService.java
@@ -1,15 +1,12 @@
 package com.fivemybab.ittabab.group.command.application.service;
 
-import com.fivemybab.ittabab.group.command.application.dto.GroupCommentDto;
-import com.fivemybab.ittabab.group.command.application.dto.GroupInfoDto;
-import com.fivemybab.ittabab.group.command.application.dto.GroupUserDto;
-import com.fivemybab.ittabab.group.command.application.mapper.GroupCommentMapper;
-import com.fivemybab.ittabab.group.command.application.mapper.GroupInfoMapper;
-import com.fivemybab.ittabab.group.command.application.mapper.GroupUserMapper;
-import com.fivemybab.ittabab.group.command.application.repository.GroupInfoRepository;
-import com.fivemybab.ittabab.group.command.application.repository.GroupUserRepository;
 import com.fivemybab.ittabab.group.command.domain.aggregate.GroupInfo;
 import com.fivemybab.ittabab.group.command.domain.aggregate.GroupUser;
+import com.fivemybab.ittabab.group.command.domain.repository.GroupInfoRepository;
+import com.fivemybab.ittabab.group.command.domain.repository.GroupUserRepository;
+import com.fivemybab.ittabab.group.query.Service.GroupQueryService;
+import com.fivemybab.ittabab.group.query.dto.GroupInfoDto;
+import com.fivemybab.ittabab.group.query.dto.GroupUserDto;
 import jakarta.transaction.Transactional;
 import lombok.RequiredArgsConstructor;
 import org.modelmapper.ModelMapper;
@@ -21,25 +18,13 @@ import java.util.List;
 @RequiredArgsConstructor
 public class GroupService {
 
-    private final GroupInfoMapper groupInfoMapper;
-    private final GroupUserMapper groupUserMapper;
-    private final GroupCommentMapper groupCommentMapper;
     private final GroupInfoRepository groupInfoRepository;
     private final GroupUserRepository groupUserRepository;
     private final ModelMapper modelMapper;
-
-    public List<GroupInfoDto> findGroupByGroupStatus(String loginId) {
-        return groupInfoMapper.findGroupByGroupStatus(loginId);
-    }
+    private final GroupQueryService groupQueryService;
 
     public GroupInfoDto findGroupByGroupId(Long groupId) {
-        return groupInfoMapper.findGroupByGroupId(groupId);
-    }
-
-    public List<GroupCommentDto> findGroupCommentsByGroupId(Long groupId) {
-        List<GroupCommentDto> commentList = groupCommentMapper.findGroupCommentsByGroupId(groupId);
-
-        return commentList;
+        return groupQueryService.findGroupByGroupId(groupId);
     }
 
     /* 모임 등록 */
@@ -50,7 +35,7 @@ public class GroupService {
 
     /* 로그인 ID -> 유저 ID */
     public Long loginIdToUserId(String loginUserLoginId) {
-        return groupInfoMapper.findUserIdByLoginId(loginUserLoginId);
+        return groupQueryService.loginIdToUserId(loginUserLoginId);
     }
 
     /* 모임 삭제 */
@@ -61,7 +46,7 @@ public class GroupService {
 
     /* 모임에 가입된 사용자 아이디 가져오는 메소드 */
     public List<Long> findGroupUserByGroupId(Long groupId) {
-        return groupUserMapper.findUserByGroupId(groupId);
+        return groupQueryService.findGroupUserByGroupId(groupId);
     }
 
     /* 모임에 신규 사용자 가입 메소드 */
