+++ src/main/java/com/fivemybab/ittabab/group/command/domain/repository/GroupInfoRepository.java
@@ -1,4 +1,4 @@
-package com.fivemybab.ittabab.group.command.application.repository;
+package com.fivemybab.ittabab.group.command.domain.repository;
 
 
 import com.fivemybab.ittabab.group.command.domain.aggregate.GroupInfo;
