+++ src/main/java/com/fivemybab/ittabab/group/command/domain/repository/GroupUserRepository.java
@@ -1,4 +1,4 @@
-package com.fivemybab.ittabab.group.command.application.repository;
+package com.fivemybab.ittabab.group.command.domain.repository;
 
 import com.fivemybab.ittabab.group.command.domain.aggregate.GroupUser;
 import org.springframework.data.jpa.repository.JpaRepository;
