+++ src/main/java/com/fivemybab/ittabab/group/query/mapper/GroupCommentMapper.java
@@ -1,6 +1,6 @@
-package com.fivemybab.ittabab.group.command.application.mapper;
+package com.fivemybab.ittabab.group.query.mapper;
 
-import com.fivemybab.ittabab.group.command.application.dto.GroupCommentDto;
+import com.fivemybab.ittabab.group.query.dto.GroupCommentDto;
 import org.apache.ibatis.annotations.Mapper;
 
 import java.util.List;
