+++ src/main/resources/mapper/group/GroupInfoMapper.xml
@@ -2,9 +2,9 @@
 <!DOCTYPE mapper
         PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN"
         "https://mybatis.org/dtd/mybatis-3-mapper.dtd">
-<mapper namespace="com.fivemybab.ittabab.group.command.application.mapper.GroupInfoMapper">
+<mapper namespace="com.fivemybab.ittabab.group.query.mapper.GroupInfoMapper">
     <select id="findGroupByGroupStatus"
-            resultType="com.fivemybab.ittabab.group.command.application.dto.GroupInfoDto">
+            resultType="com.fivemybab.ittabab.group.query.dto.GroupInfoDto">
         SELECT
         grp.group_id,
         grp.group_category_id,
@@ -23,7 +23,7 @@
     </select>
 
     <select id="findGroupByGroupId"
-            resultType="com.fivemybab.ittabab.group.command.application.dto.GroupInfoDto">
+            resultType="com.fivemybab.ittabab.group.query.dto.GroupInfoDto">
         SELECT *
         FROM group_info
         WHERE group_id = #{groupId}
