+++ src/main/resources/mapper/group/GroupUserMapper.xml
@@ -2,7 +2,7 @@
 <!DOCTYPE mapper
         PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN"
         "https://mybatis.org/dtd/mybatis-3-mapper.dtd">
-<mapper namespace="com.fivemybab.ittabab.group.command.application.mapper.GroupUserMapper">
+<mapper namespace="com.fivemybab.ittabab.group.query.mapper.GroupUserMapper">
     <select id="findUserByGroupId"
             resultType="long">
         SELECT user_id
